Fixed return value of copy_auxiliary_data_from_minc2_file() function.
The function's comments say "RETURNS : VIO_OK or VIO_ERROR" but it would sometimes return -1, which I'm assuming should be VIO_ERROR.

Fixes compiler warning about -1 not being one of the VIO_Status enum values.  Indeed, that enum has no negative values.

This could be dangerous if clients of the function are depending on sometimes getting -1 though.  On the other hand, the current situation is also dangerous if clients are depending on the docs being right!