COMP: Add enumerated value for MI_ORIGINAL_TYPE to mitype_t
The function "nc_type_to_minc2_type" returns a value of type "mitype_t".
As this function can return the value "MI_ORIGINAL_TYPE" (which is 0)
and "mitype_t" had no enumerated value of 0, a warning was being produced.

This commit adds an enumerated value of 0 to "mitype_t", allowing variables
of this type to legally hold the value "MI_ORIGINAL_TYPE".